requirements.txt
elasticsearch
PyMemoize
redis
umsgpack
gevent
